backend/requirements.txt
# Core Django dependencies
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter (optional)
django-extensions>=3.2.0

# Development and testing
django-debug-toolbar>=4.0.0
pytest>=7.0.0
pytest-django>=4.5.0
factory-boy>=3.2.0

# Code quality and formatting
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
pre-commit>=3.0.0

# Documentation
django-silk>=5.0.0  # Performance profiling (optional)

# Deployment
gunicorn>=20.1.0
whitenoise>=6.4.0

# Utilities
python-decouple>=3.8  # Environment variables
pillow>=9.5.0  # Image processing
celery>=5.2.0  # Task queue (optional)
redis>=4.5.0  # Caching and message broker (optional)
cryptography>=41.0.0  # Field-level encryption
django-cryptography  # Django encrypted fields (optional)